Add hacking rule to prevent log translations
The i18n team has decided not to translate the logs because it seems
like it not very useful; operators prefer to have them in English so
that they can search for those strings on the internet.

Since log translations have been removed completely, we should add
hacking rule to prevent future mistakes.

Change-Id: Id407764284bf967001357f303d17f9e171678120