Fix order of socket.error for catching exceptions
When an exception occurs, Python will search for the first exception
clause which matches the exception type that occurred. It doesn’t need
to be an exact match. If the exception clause represents a base class of
the raised exception, then Python considers that exception clause to be
a match [1].

Because socket.error is subclass of IOError, it should be handled before
its superclass.

[1] https://docs.quantifiedcode.com/python-anti-patterns/correctness/bad_except_clauses_order.html

Change-Id: If9070120c1d03e7eeb6d5d38ebb408fa26044274